Reworkd TermModeRestore.
Do not depend on automatically detecting when to enter terminal mode or
normal mode. Instead, let the user control this with TermModeSave
command. If never used, on BufEnter, startinsert will always be called.
This is the most common use case.

If the user wants to prevent that from happening, he has to call
TermModeSave, which will disable that behavior. Calling TermModeSave
again will re-enable it.

In the future, if there's an event for entering terminal mode, we could
make things more automatic:
   - always startinsert on BufEnter, except if a flag is set
   - set the flag when the user leaves terminal mode with C-h
   - clear the flag when going again to terminal mode

Right now we don't have a way to clear that flag when the user goes back
to terminal mode, so lets keep things manual.